requirements.txt
# Usage: pip install -r requirements.txt

gitpython>=3.1.30
opencv-python>=4.1.1
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
tqdm>=4.64.0
torch>=1.7.0
torchvision>=0.8.1
mysql-connector-python>=8.0.32
tensorboard>=2.4.1
seaborn>=0.11.0
setuptools>=65.5.1
altgraph==0.17.3
anaconda==0.0.1.1
autopep8==2.0.2
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pefile==2023.2.7
Pillow==9.4.0
pycodestyle==2.10.0
pyinstaller==5.9.0
pyinstaller-hooks-contrib==2023.1
pyparsing==3.0.9
PySimpleGUI==4.60.4
python-dateutil==2.8.2
pytz==2022.7.1
pywin32-ctypes==0.2.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl
tomli==2.0.1
typing_extensions==4.5.0
zipp==3.15.0
PySimpleGUI==4.60.4
PyAutoGUI==0.9.53
pywin32==306
Flask==2.3.2
ipython==8.13.2